Issue-315: Processing Queue ID as a long value
Queue ID retured from Jenkins can be a Long value. This is due to a recent update which resulted in the queue IDs returning a random long values instead of sequential integers. The jenkins-rest API expects this to be an Integer value. As a result, calls to jobsApi().buildWithParameters and then attempting to retrieve the queue ID will fail/return null.

This PR fixes the JobsApi and QueueApi to deal with the new long values.

Please refer to issue [315](https://github.com/cdancy/jenkins-rest/issues/315)